main up

+++ libft/ft_libc/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: flcarval <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/22 12:35:32 by flcarval          #+#    #+#             */
-/*   Updated: 2021/11/22 12:35:33 by flcarval         ###   ########.fr       */
+/*   Updated: 2021/11/24 15:01:46 by flcarval         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../header/libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
